RENAME: network_manager to client_manager
